requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
annoy==1.17.3
anyio==4.4.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.33.7
botocore==1.33.13
bs4==0.0.1
cachetools==5.5.0
certifi==2022.12.7
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
faiss-cpu==1.7.4
fastapi==0.114.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.66.1
h11==0.16.0
h5py==3.11.0
# huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
keras==3.7.0
keras-bert==0.89.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.51.0
keras-transformer==0.40.0
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.5.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numba==0.58.1
numpy==1.26.4
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==9.1.0.70
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.20.5
# nvidia-nvjitlink-cu12==12.6.68
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.8
psutil==5.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.0
pydantic_core==2.23.2
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.8.2
safetensors==0.4.4
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
# starlette==0.40.0
# sympy==1.13.2
# tensorboard==2.8.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.9.3
tensorflow-macos
tensorflow-metal
# tensorflow-estimator==2.8.0
# tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
# tokenizers==0.19.1
torch==2.7.1
# torchvision==0.19.0
tqdm==4.66.5
transformers==4.53.0
# triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
# urllib3==2.5.0
usearch==2.15.1
uvicorn==0.30.6
waitress==3.0.1
Werkzeug==2.2.3
# wrapt==1.16.0
zipp==3.20.1
